+++ src/parse/is_expected.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   is_expect_op.c                                     :+:      :+:    :+:   */
+/*   is_expected.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ikawamuk <ikawamuk@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/03 16:17:26 by ikawamuk          #+#    #+#             */
-/*   Updated: 2026/02/14 15:28:09 by ikawamuk         ###   ########.fr       */
+/*   Updated: 2026/02/28 07:05:57 by ikawamuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 bool	is_expected(const char *op, t_token *token)
 {
 	return (token->type == TK_RESERVED
-		&& strlen(op) == token->len
+		&& strlen(op) == (size_t)token->len
 		&& !memcmp(op, token->str, token->len)
 	);
 }
