+++ src/parse/new_node/new_func_call_leaf.c
@@ -6,7 +6,7 @@
 /*   By: ikawamuk <ikawamuk@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/15 05:11:27 by ikawamuk          #+#    #+#             */
-/*   Updated: 2026/02/28 06:55:32 by ikawamuk         ###   ########.fr       */
+/*   Updated: 2026/02/28 07:04:18 by ikawamuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 void	*xaalloc(size_t size);
 
-t_tree	*new_func_call_leaf(const char *func_name, size_t len, t_tree *args)
+t_tree	*new_func_call_leaf(const char *func_name, int len, t_tree *args)
 {
 	t_tree	*new = xaalloc(sizeof(t_tree));
 	new->type = ND_FUNC_CALL;
